+++ src/expr/sql.rs
@@ -1,6 +1,7 @@
 //! Convert Expr into ast::Expr
 use crate::{
     ast,
+    data_type::DataType,
     expr::{self, Expr},
     visitor::Acceptor,
 };
@@ -176,7 +177,9 @@ impl<'a> expr::Visitor<'a, ast::Expr> for FromExprVisitor {
             | expr::function::Function::CharLength
             | expr::function::Function::Lower
             | expr::function::Function::Upper
-            | expr::function::Function::Random(_) => ast::Expr::Function(ast::Function {
+            | expr::function::Function::Random(_)
+            | expr::function::Function::Least
+            | expr::function::Function::Greatest => ast::Expr::Function(ast::Function {
                 name: ast::ObjectName(vec![ast::Ident::new(function.to_string())]),
                 args: arguments
                     .into_iter()
@@ -209,10 +212,13 @@ impl<'a> expr::Visitor<'a, ast::Expr> for FromExprVisitor {
                 expr: arguments[0].clone().into(),
                 r#in: arguments[1].clone().into(),
             },
-            expr::function::Function::Cast(into) => ast::Expr::Cast {
+            expr::function::Function::CastAsText => ast::Expr::Cast {
                 expr: arguments[0].clone().into(),
-                data_type:  into.clone().into(),
+                data_type:  DataType::text().into(),
             },
+            expr::function::Function::CastAsFloat => todo!(),
+            expr::function::Function::CastAsInteger => todo!(),
+            expr::function::Function::CastAsDateTime => todo!(),
         }
     }
     // TODO implement this properly