+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Internal code uses `Relation.name()` for table addressing but user facing functions may use `Table.path()` (ie sql addressing)
 - Renamed bivariate_min and bivariate_max to least and greatest
 - Cast to string before MD5 for protection
+- Implemented `least` and `greatest` (untested)
 
 ## [0.2.2] - 2023-08-29
 ### Changed